misc: Mv `StrawberryGQLContext` type
